api/requirements.txt
amqp==2.6.1
appdirs==1.4.4
asgiref==3.2.10
attrs==20.1.0
billiard==3.6.3.0
black==19.10b0
celery==4.4.7
certifi==2020.6.20
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
Django==3.1
djangorestframework==3.11.1
factory-boy==3.0.1
Faker==4.1.2
filelock==3.0.12
identify==1.4.28
idna==2.10
importlib-metadata==1.7.0
importlib-resources==3.0.0
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.11
Markdown==3.2.2
MarkupSafe==1.1.1
nodeenv==1.4.0
pathspec==0.8.0
pre-commit==2.6.0
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
regex==2020.7.14
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
virtualenv==20.0.31
zipp==3.1.0
